backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
pdfplumber>=0.10.0
docling>=2.0.0
easyocr>=1.7.0
Pillow>=10.0.0
python-docx>=1.0.0
pytesseract>=0.3.10
aiosqlite>=0.19.0
python-dotenv>=1.0.0
sentence-transformers==2.7.0
chromadb==0.5.0
google-generativeai==0.7.0
